Render error page when section load fails
Fixes the code that renders an error page when loading a section module (i.e., its webpack chunk)
fails with error. There was a bug in that code: it set `context.primary` to a React element with
the error page UI, but didn't call the code needed to actually render that UI.

The `SET_SECTION` dispatch is needed to reset flags like `state.ui.section` and `state.ui.hasSidebar`.
These flags are used by `Layout` and affect rendering of the error page. For example, if we didn't
reset the `hasSidebar` flag, the error page would render a sidebar area if the last loaded section
supported it.